rpiserver/pyproject.toml
[tool.poetry]
name = "rpiserver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
"RPi.GPIO" = {version = "^0.7.0", optional = true}
rpi-gpio-emu = {version = "^0.0.1", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.extras]
rpi-gpio-real = ["RPi.GPIO"]
rpi-gpio-emu = ["rpi-gpio-emu"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
